(#7656) Add generate action to module_tool face
Before this patch, the module_tool face did not have the ability to
generate boilerplate for a new module, and there was no test coverage
for the module_tool generate action.

This patch adds support for generating boilerplate for a new module to
the module_tool face.

Example usage:

    $ puppet module_tool generate username-modulename

This patch includes unit tests for the module_tool generate action.