(#7656) Add install action to module_tool face
Before this patch, the module_tool face did not have the ability to
install a module from a repository or release archive, and there was no
test coverage for the module_tool install action.

This patch adds support for installing a module from a repository or
release archive to the module_tool face.

Example usage:

    # Install a module from the default repository:
    $ puppet module_tool install username-modulename

    # Install a specific module version from a repository:
    $ puppet module_tool install username-modulename --version=0.0.1

    # Install a module into a specific directory:
    $ puppet module_tool install username-modulename --install-dir=path

    # Install a module from a release archive:
    $ puppet module_tool install username-modulename-0.0.1.tar.gz

This patch includes unit tests for the module_tool install action.