(#7656) Add search action to module_tool face
Before this patch, the module_tool face did not have the ability to
search a repository for a module, and there was no test coverage for the
module_tool search action.

This patch adds support for searching a repository for a module to the
module_tool face.

Example usage:

    $ puppet module_tool search modulename

This patch includes unit tests for the module_tool search action.