(#7656) Refactor the module face build action
Before this patch, the build action attempted to located the root path
of the module being built before calling the
`Puppet::Module::Tool::Applications::Builder#run` method.

This patch moves the logic of locating the module root path to the
`Puppet::Module::Tool::Applications::Builder` class, as the root path of
the module would never be changed directly in the faces code.