(#7656) Rename contact to make_http_request
Before this patch, the repository module used method names such as
`contact` to refer the action of making an http request, and
`read_contact` to refer to the action of reading an http response. This
caused readability issues.

This patch fixes these issues by renaming the following methods:

 * contact -> make_http_request
 * read_contact -> read_response

This patch also updates the related unit tests to reflect the new names
of these methods.