(#7656) Use rspec 2.x in generator templates
Before this patch, the generator action generated rspec templates setup
to be run with rspec 1.x, and assumed the usage of rubygems. Rspec 2.x
is now the preferred version and should be used instead.

This patch fixes this issue by updating the rspec templates to use rspec
2.x and removed the assumption of rubygems. With the moved to rspec 2.x,
the `spec.opts` template is no longer required, and is being removed.