(#7656) Rename full_name to full_module_name
Before this patch, the module application uses the term full_name when
referring to a "full" module name (username-modulename). The use of
full_name causes a readability issue, as the name does not represent
what it's describing.

This patch fixes this issue by renaming full_name to full_module_name