(#7656) Rename `Modulefile` to `ModulefileReader`
Before this patch, the modulefile module uses a class name of
`Modulefile` to refer to a class which reads modulefiles. The use of
`Modulefile` in this case is a bit misleading and causes a readability
issue.

This patch fixes this issue by renaming `Modulefile` to
`ModulefileReader`, which better describes the behavior of the class.