(#7656) Update comment string in the metadata module
Before this patch, the metadata module has a comment which asks the
question "Shouldn't this be it's own class?", when referring to the how
the type metadata is managed.

This has been discussed during a recent code review, and it has been
decided that the answer to this question is no. Simple types(eg: Hash)
are OK in this context. This patch removes this question as it is no
longer valid.