(#7656) Add details about module testing to templates
Before this patch, the `init.pp.erb` test template does not contain any
information on how a user is supposed to use the tests, there is nothing
guiding the user on what their next step should be.

This patch solves this problem by adding a brief summary on how module
testing works, and how to use them. A link to the online documentation
is also provided as a convenience to the user.