(#7656) Rename module_tool face to module
Before this patch the module face was named module_tool, this is not a
very attractive name. Look at it, you have to type an underscore just to
use it.

This patch fixes this issue by simply renaming the face from module_tool
to module. The patch does not change the name of the module tool library
files, doing so would cause a conflict on the existing
`lib/puppet/module.rb` module.