Fixed #11263 - Update regex for metadata_expire in yumrepo
The current metadata_expire attribute only takes time in seconds.
This is not correct behaviour and limits the ability of the yumrepo
type to manage metadata.  The correct inputs to this attribute should
be:

* Time in seconds
* Integer suffixed with d, h, m for days, hours and minutes
* "never" indicating that metadata should never expire

This patch updates the attributes regex to match this behaviour.

There are no test updates because there are no current Yumrepo RSpec
tests.