(#11198) Modulefiles should have a default license
Before this patch, the module generate action generates Modulefiles
with the license set to UNKNOWN, which is not a valid license.

This patch fixes this issue by setting the default value of the
`Puppet::Module::Tool::Metadata#license` attribute to 'Apache License,
Version 2.0' by default. This patch also adds a new unit test verifying
this this change.