(#11276) Mark module tool tests as failing on Windows
We're currently getting failures like:

   Failure/Error:
   Puppet::Module::Tool::Applications::Builder.run(@full_module_name)
     RuntimeError:
       Could not create myuser-mymodule-0.0.1.tar

Until we can deal with the module tool's reliance on tar, either by
recommending a tar installation for windows or not needing to, the
module tool won't work on windows and I'm marking the specs accordingly.

`puppet module search` currently works on windows, but it's one of the
few actions that doesn't deal with tar files.