(#11246) Add README_DEVELOPER describing UTF-8 in Puppet
Without this patch, developers of Puppet don't have a clear place to get
a high level understanding of the way other Puppet developers are
working with UTF-8 and the differences in character encodings between
Ruby 1.8 and 1.9.

This patch addresses this problem by adding a new document,
README_DEVELOPER.md where developers and contributors can look to for
high level information.