Check for bad plist doctype strings
When a plist file has a bad doctype string it will still pass lint,
but ruby's libxml will flag it as malformed and throw an error. This
is compounded by the fact that Apple ships a plist in 10.6 that has
a bad doctype string (/System/Library/LaunchDaemons/org.ntp.ntpd.plist).
It's not our job to 'correct' the file, so instead a debug message is
thrown, the plist file is read into a variable, and the doctype string
in the variable is substituted for a correct doctype string.