(#10109) Make resourcefile work with composite namevars
The code for creating the resourcefile was directly calling
resource.name_var, which was causing problems with resources that have
composite namevars (since, for these, Type#name_var will return false).
This patch sanitizes the process by first checking whether there is a
single namevar, and simply calling resource.ref if there is not one.

Reviewed-by: Matt Robinson