Allow failback to facter/util/plist
The CFPropertyList library vendored in facter/util will parse plists faster
than the facter/util/plist library, however if CFPropertyList isn't installed
there needs to be a failback to facter/util/plist.  

If facter/util/cfpropertylist is installed, we will default to that behavior,
but if it isn't, the $legacy_plist_support variable will be set to true and
Puppet will default back to facter/util/plist library.