Update tests to align with $legacy_plist_support
We've introduced facter/util/cfpropertylist as a method by which to
read/save plist files, so we need to test for that. This commit
adapts our tests to ultimately test for the same end-result while
adjusting what Mocha expects depending on the plist library used.

The $legacy_plist_support variable will be set in the launchd
provider when we're using the older facter/util/plist library, so
we're using the presence of that variable to indicate whether we're
using the older facter/util/plist library or newer
facter/util/cfpropertylist library.