Add :setup task to populate variables
There were problems with using `git describe` within a global
scope to set the @version variable (namely that our Jenkins jobs
would try and evaluate this with custom tags that don't match our
release tags).  To remedy this, the class instance variables are
now getting set in a :setup task that is getting called by the
package:apple tasks. The Jenkins jobs never attempt to run this
task and the build doesn't fail (yay!). Doing this also preserves
the 'rc' tags that we desired.