Access user password hash in OS X 10.7
A first attempt at accessing the user's password hash in 10.7 by
inspecting the user's plist in
/var/db/dslocal/nodes/Default/users/test.plist. This method requires
saving a binary plist to /tmp/username.plist in order to use plutil
-convert xml1 to convert it to xml (as Plist::parse_xml can only read
XML plists).

Use Puppet::Util::Package.versioncmp()

Because we're comparing OS X versions that are ultimately strings,
we need a method to catch the scenario where '10.10' < '10.7'.
Puppet::Util::Package.versioncmp() does this well, so we will use
it in favor of simple string comparison.