Fix Ruby 1.8.5-incompatible code in spec setup
The Range#step method in Ruby 1.8.5 expects to be passed a block, to
which it will yield each stepped element. In 1.8.7, it will simply
return an Enumerator (which we then enumerate) if no block is provided,
but in 1.8.5, it will try to yield anyway and fail because there
is no block.

Reviewed-By: Josh Cooper <josh@puppetlabs.com>