(#11404) Move markdown_header and markdown_definitionlist to Puppet::Util::Docs
The markdown_header and markdown_definitionlist methods are needed in places
other than the Puppet::Util::Reference class. In particular, they may become
necessary when assembling fragments of documentation from other fragments,
much as we need to do for the :provider parameter.

This commit puts them in the Puppet::Util::Docs module, which leaves them
available when generating the references but also exposes them in many other
places.