Maint: Make indentation for markdown definition lists more readable and reliable
Previously, we indented subsequent lines of definition lists with a four-space
hanging indent. This ensured that multi-paragraph DD elements would work, but
was unsightly. This commit changes the hanging indent for definitions to 2
spaces, which also works but is much more legible before being processed into HTML.

This commit also adds sanitization (via the scrub method) for definition text.
Since scrub() is idempotent, and since the consequences of text with
unreliable indentation making it into a DD element are so annoying, it makes
sense to hedge our bets when creating the element.

Test update included.