(#11404) Fix broken and unreliable indentation in provider lists in type reference
Previously, portions of a provider's documentation string could be wildly
outdented relative to their surroundings, due to the interaction of some
inaccurate assumptions; see http://projects.puppetlabs.com/issues/11404 for
more details.

This commit fixes these formatting glitches by normalizing the indentation of
the various fragments we assemble into the provider parameter's doc string. We
normalize the parameter's @doc explicitly with scrub(), and then we render the
list of providers as a definition list, which implicitly normalizes indentation.