Maint: Padding should be added when dochook strings are consumed, not hardcoded into them
Previously, the return values of the dochook_ methods in provider.rb had a
leading two-space padding hardcoded into them. This was mostly fine, but had
the potential to interact badly with semantic Markdown indentation. This
commit moves the padding into the method where we compose those strings into
their final position, which gets us some better separation of content and
formatting.

This commit also changes our display of extra info by giving it its own
paragraph. Previous behavior was undefined and depended on whether the doc string
ended with more than one newline; it usually appended the extra info to the
final paragraph of the documentation.