Maint: Fix redhatfedorasusecentosslesoelovm in type reference
Many providers are the default for an array of operatingsystem values.
Although the dochook(:defaults) method was handling multiple defaults already,
it wasn't handling multiple values for a single default very well, so we would
say a provider was default for operatingsystem == redhatfedorasusecentosslesoelovm.

This commit refactors that to render arrays of values for a single default as
a comma-separated list, without breaking the way we handle string values for a
single default.