(#11423) Better error when imported resources overlap.
When multiple nodes export a resource with a duplicate type and title, and you
import both, the error message misleadingly tells you that this was a *local*
duplication.

This improves the error message to cover all the bases, which is suboptimal -
it should tell you where the duplicate comes from - but at least doesn't lead
to a wild goose hunt for the *local* duplicate that doesn't exist.

Signed-off-by: Daniel Pittman <daniel@puppetlabs.com>