(#10586) Don't copy owner and group when sourcing files from master
Previously, puppet on Windows was not able to source files from the
master, because it was attempting to translate the uid/gid from
the Unix master into a Windows account, and obviously failing.

This commit skips the owner and group properties when copying them
from non-local sources, i.e. sources whose URIs have a 'puppet'
scheme.

If the source comes from a local source, then puppet behaves the same
as it did previously, it copies the owner and group if the source
volume supports Windows ACLs, e.g. C:/, samba mapped drives, or uses
default values if the volume does not, e.g. VMware shared drives.