(#11408) Fix fact and plugin sync on Windows
Previously, fact and pluginsync were broken on Windows, because it was
defaulting the owner and group to Process.uid/gid, and then failing to
translate them into Windows SIDs.

This commit changes the default file owner to the current user name,
and the default file group to Nobody, which is the group that Windows
typically applies to newly created files.