(#11414) Save/execute changes on versions of Augeas < 0.3.6
Versions of Augeas prior to 0.3.6 didn't report their version number, so a
fallback of executing changes once in need_to_run? and again in execute_changes
is performed.  Otherwise a save is done in need_to_run? and this is re-used in
execute_changes.

The /augeas/events/saved node is used to tell whether the latter optimisation
happened, but the return value of #match wasn't tested correctly (it's an empty
array).