(#10111) - OCSP certificate verification
This commit adds the foundation layer of an OCSP responder for Puppet.
It also includes a puppet face to verify a given certificate
status remotely through rest.

The verification of the certificate status is done with the actual
CRL.

Note: OCSP is a pure request/response protocol. Puppet doesn't really
support such model in its REST handling. So to almost achieve this,
we "save" Request to the REST indirector, and returns then the
OCSP response. Unfortunately Puppet insists to return YAML content,
which kind of make us not compatible with other OCSP responder.

Signed-off-by: Brice Figureau <brice-puppet@daysofwonder.com>