Refactor: configurer now returns report.exit_status
None of the configurer client was directly using the returned
report, but only its exit_status.
Since now the configurer might run in a child process, we can't return
the report, only a status code.
This commit changes the configurer and call site so that only the report
exit code is returned.

Signed-off-by: Brice Figureau <brice-puppet@daysofwonder.com>