Disable SRV lookups of hosts by default when testing
The old behavior of leaving SRV lookups turned on by default in tests
was causing problems if the machine running the tests reported itself
as having a domain name that has SRV records setup for Puppet.  This
would cause extra calls to Puppet.warning due to the failed connection
attempts, which could cause a problem if the test was looking
for specific warnings (which
spec/unit/indirector/facts/inventory_service_spec.rb does).