(#4836) - Agent --disable should allow to put a message
This way it is possible to let other users know why a puppet agent
has been disabled.

Usage:
puppet agent --disable "because working on backup"

When later on, the puppet agent run would abort with:
"Run of Puppet configuration client is administratively
disabled because working on backup; skipping"

It's also possible to not set a message, in which case a default
message is provided.

Signed-off-by: Brice Figureau <brice-puppet@daysofwonder.com>