Use all lower-case file name for Puppet::Util::Instrumentation::Instrumentable
Ruby convention is to use all lower-case when naming files &
directories of files, and to require them using the lower-case name.

While the file lib/puppet/util/instrumentation/Instrumentable.rb could
be required using the lower-case convention on a case-preserving but
-insensitive file system, the file could not be found if the same
require was used on a file system that was both case-preserving and
case-sensitive.