(#11595) Delete unused activerecord catalog find
The find method on lib/puppet/indirector/catalog/active_record.rb was
written in such a way that it could never be used.

First, you had to pass cache_integration_hack as an option to get it to
return anything except nil, and this wasn't used in the code.

Second, if you ever did get past the hack option,
resource.to_transportable was called even though that method was never
defined anywhere.

We noticed this when looking for code that had been orphaned by #11552
to delete TransObjects and TransBuckets.  There was a to_trans method
that was deleted, but never was there a to_transportable.

Paired-with: Patrick Carlisle <patrick@puppetlabs.com>