(#11499) Better validation for IPv4 and IPv6 address in host type.
The previous code was fairly lax in validation, and would allow a bunch of
invalid addresses through - as well as rejecting some legal, but uncommon,
IPv6 address types.

This adds substantial testing, especially around IPv6 addressing, and replaces
the older validation with new, fancy stuff that works for all the cases.

Signed-off-by: Daniel Pittman <daniel@puppetlabs.com>