(#11552) Fix `puppet resource` ability to display parameters
The original commit for #11552 removing TransObject and TransBucket
removed the code that manually added parameters when converting from a
Puppet::Type::Resource to a Resource.  No tests in Puppet caught this
problem, but a test of the puppetral MCollective provider caught that
Puppet::Resource could no longer give back paramger information.

This restores the ability to retrieve parameters and adds tests to
Puppet.  Long term the logic for adding parameters really shouldn't live
in a method that converts between resource representations, but for now
I'm fixing the regression to work the same way it used to.