(#11740) Disable failing test on Windows
Jenkins uncovered a race condition on Windows when a parent process
tries to get the exit code of its child process. If the child exits
before the parent calls OpenProcessToken, then the parent cannot get
the child's exit code.

This issue has been present for a long time, and isn't trival to fix,
because the spec tests assume waitpid2 behaves consistently on POSIX
and Windows.

I'm disabling the test on Windows for now, and will revert this change
when I have a real fix for #11740.

Reviewed-By: Jeff McCune <jeff@puppetlabs.com>