Revert "Access user password hash in OS X 10.7"
This reverts commit 11b8c5ad3fb99263c8484fecc767b69cd80d3156.

This needs to be reverted because I missed the use of backticks (``),
effectively %x{}, in the original pull request and this is causing a big
problem for the CI tests on non Mac OS X systems that don't have plutil.

Gary and I are going to re factor the pull request to make the command
calls easier to mock out on non Mac OS X platforms.  We can't do this
quickly tonight though.

Updated information will be in ticket #11293