(#11714) Use `%~dp0` to resolve bat file's install directory
This commit uses the `%~dp0` batch script modifier to resolve the
drive and path of the directory containing the envpuppet.bat
file. This eliminates the need for hard coded paths within the script
itselfIt also uses `%VAR:\=/%` to substitute each backslash for a
forward slash in the RUBYLIB environment

Also added a section about running the spec tests on Windows.

Reviewed-By: Jeff McCune <jeff@puppetlabs.com>