Maint: Don't use cached catalogs
Previously, the agent would attempt to connect to the master, and fail
due to the SSL 'hostname did not match' error, but then it could use
an empty cached catalog from a different test and return exit code 0,
as nothing had changed.

This commit ensures we don't use a cached catalog, and therefore will
exit with a non-zero status as expected.

Paired-With: Jeff McCune <jeff@puppetlabs.com>