Maint: Actually test that a file can be sourced from remote master
Previously, the test was redirecting the manifest to the wrong file,
so the master was serving an empty catalog to agents. This change was
made in 803e92f, however, it did not actually cause an acceptance test
failure because of bug #11764 in the acceptance test framework
itself. The framework was fixed in commit b8cda1d, merged into master
1/5/12, and the next run of the acceptance tests in the morning of
1/6/12 is when the failure started.

This commit changes the test to write the manifest to the right
file. Also the commit requires puppet agent to return exit code 2,
because we expect the agent to make changes and we are using
`--detailed-exit-codes` (which is automatically set when using
`--test`).

Paired-with: Jeff McCune <jeff@puppetlabs.com>